+++ includes/class-dashboard.php
@@ -26,6 +26,8 @@ class Dashboard {
 		add_action( 'wp_dashboard_setup', [ self::class, 'remove_dashboard_widgets' ] );
 		add_action( 'wp_network_dashboard_setup', [ self::class, 'remove_dashboard_widgets' ] );
 		add_action( 'wp_user_dashboard_setup', [ self::class, 'remove_dashboard_widgets' ] );
+
+		add_action( 'wp_dashboard_setup', [ self::class, 'add_bm_academy_dashboard_widget' ] );
 	}
 
 	/**
@@ -48,6 +50,24 @@ class Dashboard {
 		}
 	}
 
+	public static function add_bm_academy_dashboard_widget() {
+		wp_add_dashboard_widget( 'bm_academy', __( 'Bernskiold Media Academy', 'bm-wp-experience' ), static function () {
+			$feed = fetch_feed( 'https://bernskioldmedia.com/sv/feed/' );
+
+			if ( is_wp_error( $feed ) ) {
+				echo __( 'Unfortunately the academy content is not currently available.', 'bm-wp-experience' );
+			} else {
+				$feed->init();
+				$feed->set_output_encoding( 'UTF-8' );
+				$feed->handle_content_type();
+				$feed->set_cache_duration( 21600 );
+				$limit = $feed->get_item_quantity( 5 );
+				$items = $feed->get_items( 0, $limit );
+
+				include BM_WP_Experience::get_view_path( 'dashboard-widgets/academy-feed' );
+			}
+		}, null, null, 'side' );
+	}
 }
 
 Dashboard::init();